Implement v8::Object::[GS]et() for numeric keys.
Make the get/set operation a trait that is implemented for u32 and
ValueT.

Coincidentally fixes a value comparison bug.  A Value is a pointer to a
pointer.  The first one points to the HandleScope slot, the second one
to the actual JS value.  It's legal for multiple slots to point to the
same value.  v8.rs compared the level-1 pointers but, because those are
not unique, it now compares the level-2 pointers.